requirements.txt
fastapi
httpx
pycountry
uvicorn
python-dotenv
fpdf2
asyncio
pydantic-settings